+++ src/Symfony/Component/Console/Input/ArgvInput.php
@@ -263,11 +263,11 @@ class ArgvInput extends Input
         }
 
         if (null === $value) {
-            if ($option->isValueRequired()) {
+            if (InputOption::VALUE_REQUIRED === $option->valueMode()) {
                 throw new RuntimeException(\sprintf('The "--%s" option requires a value.', $name));
             }
 
-            if (!$option->isArray() && !$option->isValueOptional()) {
+            if (!$option->isArray() && InputOption::VALUE_OPTIONAL !== $option->valueMode()) {
                 $value = true;
             }
         }