+++ src/Symfony/Component/Console/Input/ArrayInput.php
@@ -224,11 +224,11 @@ class ArrayInput extends Input
         $option = $this->definition->getOption($name);
 
         if (null === $value) {
-            if ($option->isValueRequired()) {
+            if (InputOption::VALUE_REQUIRED === $option->valueMode()) {
                 throw new InvalidOptionException(\sprintf('The "--%s" option requires a value.', $name));
             }
 
-            if (!$option->isValueOptional()) {
+            if (InputOption::VALUE_OPTIONAL !== $option->valueMode()) {
                 $value = true;
             }
         }